pyproject.toml
[tool.poetry]
name = "magic_tree"
version = "0.1.0"
description = "Defines the `MagicTreeDict` class - which extends  `defaultdict` into an auto-building recursive tree for intuitive manipulation, analysis, and visualization of hierarchical data 🌱✨"
authors = ["jonmatthis <jonmatthis@gmail.com>"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7, <4"
rich = "^13.6.0"
tabulate = "^0.9.0"
pandas = "^2.1.1"
pydantic = "<2.0.0"
pyperclip = "^1.8.2"
python-dotenv = "^1.0.0"
langchain = "^0.0.315"
openai = "^0.28.1"
anthropic = "^0.3.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
bumpver = "^2023.1127"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
